more card refactor
