+++ src/containers/board/components/hand-zone.js
@@ -1,8 +1,7 @@
 import React, { useContext, useMemo } from 'react';
 import styled, { css } from 'styled-components';
 import { StoreContext } from '../../../store/context';
-// import HandCard from '../../../components/card/hand';
-import SimpleCard from '../../../components/card/simple';
+import HandCard from '../../../components/card/hand-card';
 import { CARDSTATUS } from '../../../utils/constants';
 
 const S = {};
@@ -11,7 +10,7 @@ S.Hand = styled.div`
   position:absolute;
   width:100%;
   height:20rem;
-  ${'' /* bottom:0px; */}
+  bottom:0px;
 
   background-color:black;
 `;
@@ -20,7 +19,7 @@ function Hand() {
   const { hand } = useContext(StoreContext);
 
   const cardsInHand = useMemo(() => {
-      const validCards = hand.filter(h => h.status >= CARDSTATUS.HAND && h.status <= CARDSTATUS.HAND_HOLDING).map((c, idx) => ({
+      const validCards = hand.filter(h => h.status === CARDSTATUS.HAND).map((c, idx) => ({
         ...c,
         handPosition: idx
       }));
@@ -28,11 +27,12 @@ function Hand() {
     },
     [ hand ]
   );
+  // const cardsInHand = [];
 
   return (
     <S.Hand id="hand">
       {cardsInHand.map(c => (
-        <SimpleCard data={c} key={c.cardIdx} />
+        <HandCard data={c} key={c.cardIdx} />
       ))}
     </S.Hand>
   );
